modify forms dash

+++ routes/web.php
@@ -47,20 +47,20 @@ Route::prefix('admin')->middleware('auth')->group(function () {
     Route::get('dashboard', [GeneralController::class, 'dashboard'])->name('admin.dashboard');
 
     // General settings
-    Route::get('general-settings', [GeneralController::class, 'general'])->name('admin.general');
-    Route::post('general-settings', [GeneralController::class, 'generalUpdate'])->name('admin.general.update');
+    Route::get('general-settings', [GeneralController::class, 'general'])->middleware('can:isAdmin')->name('admin.general');
+    Route::post('general-settings', [GeneralController::class, 'generalUpdate'])->middleware('can:isAdmin')->name('admin.general.update');
 
     // About
-    Route::get('about', [GeneralController::class, 'about'])->name('admin.about');
-    Route::post('about', [GeneralController::class, 'aboutUpdate'])->name('about.update');
+    Route::get('about', [GeneralController::class, 'about'])->middleware('can:isAdmin')->name('admin.about');
+    Route::post('about', [GeneralController::class, 'aboutUpdate'])->middleware('can:isAdmin')->name('about.update');
 
     // Manage Banner
-    Route::get('banner', [BannerController::class, 'index'])->name('admin.banner');
-    Route::get('banner/create', [BannerController::class, 'create'])->name('admin.banner.create');
-    Route::post('banner/create', [BannerController::class, 'store'])->name('admin.banner.store');
-    Route::get('banner/edit/{id}', [BannerController::class, 'edit'])->name('admin.banner.edit');
-    Route::post('banner/edit/{id}', [BannerController::class, 'update'])->name('admin.banner.update');
-    Route::delete('banner/destroy/{id}',[BannerController::class, 'destroy'])->name('admin.banner.destroy');
+    Route::get('banner', [BannerController::class, 'index'])->middleware('can:isAdmin')->name('admin.banner');
+    Route::get('banner/create', [BannerController::class, 'create'])->middleware('can:isAdmin')->name('admin.banner.create');
+    Route::post('banner/create', [BannerController::class, 'store'])->middleware('can:isAdmin')->name('admin.banner.store');
+    Route::get('banner/edit/{id}', [BannerController::class, 'edit'])->middleware('can:isAdmin')->name('admin.banner.edit');
+    Route::post('banner/edit/{id}', [BannerController::class, 'update'])->middleware('can:isAdmin')->name('admin.banner.update');
+    Route::delete('banner/destroy/{id}',[BannerController::class, 'destroy'])->middleware('can:isAdmin')->name('admin.banner.destroy');
 
      // Manage Portfolio Categories
      Route::get('portfolio-categories', [PcategoryController::class, 'index'])->name('admin.pcategory');
@@ -71,7 +71,7 @@ Route::prefix('admin')->middleware('auth')->group(function () {
 
      // Manage asistent
     Route::get('asistent', [AsistentController::class, 'index'])->name('admin.asistent');
-    Route::get('asistent/create/{id}', [AsistentController::class, 'create'])->name('admin.asistent.create');
+    Route::get('asistent/create', [AsistentController::class, 'create'])->name('admin.asistent.create');
     Route::post('asistent/create', [AsistentController::class, 'store'])->name('admin.asistent.store');
     Route::get('asistent/edit/{id}', [AsistentController::class, 'edit'])->name('admin.asistent.edit');
     Route::post('asistent/edit/{id}', [AsistentController::class, 'update'])->name('admin.asistent.update');
@@ -102,23 +102,15 @@ Route::prefix('admin')->middleware('auth')->group(function () {
      Route::delete('tags/destroy/{id}',[TagController::class, 'destroy'])->name('admin.tag.destroy');
 
      // Manage Blog
-    Route::get('post',[PostController::class, 'index'])->name('admin.post');
-
-    Route::get('post/create',[PostController::class, 'create'])->name('admin.post.create');
-
-    Route::post('post/create',[PostController::class, 'store'])->name('admin.post.store');
-
-    Route::get('post/edit/{id}',[PostController::class, 'edit'])->name('admin.post.edit');
-
-    Route::post('post/edit/{id}',[PostController::class, 'update'])->name('admin.post.update');
-
-    Route::get('post/trash',[PostController::class, 'trash'])->name('admin.post.trash');
-
-    Route::post('post/{id}/restore',[PostController::class, 'restore'])->name('admin.post.restore');
-
-    Route::delete('post/trash/{id}',[PostController::class, 'destroy'])->name('admin.post.destroy');
-
-    Route::delete('post/destroy/{id}',[PostController::class, 'deletePermanent'])->name('admin.post.deletePermanent');
+    Route::get('post',[PostController::class, 'index'])->middleware('can:isAdmin')->name('admin.post');
+    Route::get('post/create',[PostController::class, 'create'])->middleware('can:isAdmin')->name('admin.post.create');
+    Route::post('post/create',[PostController::class, 'store'])->middleware('can:isAdmin')->name('admin.post.store');
+    Route::get('post/edit/{id}',[PostController::class, 'edit'])->middleware('can:isAdmin')->name('admin.post.edit');
+    Route::post('post/edit/{id}',[PostController::class, 'update'])->middleware('can:isAdmin')->name('admin.post.update');
+    Route::get('post/trash',[PostController::class, 'trash'])->middleware('can:isAdmin')->name('admin.post.trash');
+    Route::post('post/{id}/restore',[PostController::class, 'restore'])->middleware('can:isAdmin')->name('admin.post.restore');
+    Route::delete('post/trash/{id}',[PostController::class, 'destroy'])->middleware('can:isAdmin')->name('admin.post.destroy');
+    Route::delete('post/destroy/{id}',[PostController::class, 'deletePermanent'])->middleware('can:isAdmin')->name('admin.post.deletePermanent');
 
     // Manage Testimonials
     Route::get('testimonials', [TestimonialController::class, 'index'])->name('admin.testi');
@@ -130,7 +122,7 @@ Route::prefix('admin')->middleware('auth')->group(function () {
 
     // Manage Pages
     Route::get('horario', [HorarioController::class, 'index'])->name('admin.horario');
-    Route::get('horario/create/{id}', [HorarioController::class, 'create'])->name('admin.horario.create');
+    Route::get('horario/create', [HorarioController::class, 'create'])->name('admin.horario.create');
     Route::post('horario/create', [HorarioController::class, 'store'])->name('admin.horario.store');
     Route::get('horario/edit/{id}', [HorarioController::class, 'edit'])->name('admin.horario.edit');
     Route::post('horario/edit/{id}', [HorarioController::class, 'update'])->name('admin.horario.update');
@@ -153,12 +145,12 @@ Route::prefix('admin')->middleware('auth')->group(function () {
     Route::delete('faq/destroy/{id}',[FaqController::class, 'destroy'])->name('admin.faq.destroy');
 
     // Manage Partners
-    Route::get('partners', [PartnerController::class, 'index'])->name('admin.partner');
-    Route::get('partners/create', [PartnerController::class, 'create'])->name('admin.partner.create');
-    Route::post('partners/create', [PartnerController::class, 'store'])->name('admin.partner.store');
-    Route::get('partners/edit/{id}', [PartnerController::class, 'edit'])->name('admin.partner.edit');
-    Route::post('partners/edit/{id}', [PartnerController::class, 'update'])->name('admin.partner.update');
-    Route::delete('partners/destroy/{id}',[PartnerController::class, 'destroy'])->name('admin.partner.destroy');
+    Route::get('partners', [PartnerController::class, 'index'])->middleware('can:isAdmin')->name('admin.partner');
+    Route::get('partners/create', [PartnerController::class, 'create'])->middleware('can:isAdmin')->name('admin.partner.create');
+    Route::post('partners/create', [PartnerController::class, 'store'])->middleware('can:isAdmin')->name('admin.partner.store');
+    Route::get('partners/edit/{id}', [PartnerController::class, 'edit'])->middleware('can:isAdmin')->name('admin.partner.edit');
+    Route::post('partners/edit/{id}', [PartnerController::class, 'update'])->middleware('can:isAdmin')->name('admin.partner.update');
+    Route::delete('partners/destroy/{id}',[PartnerController::class, 'destroy'])->middleware('can:isAdmin')->name('admin.partner.destroy');
 
     // Manage Services
     Route::get('formadores', [FormadoresController::class, 'index'])->name('admin.formadores');
@@ -177,19 +169,19 @@ Route::prefix('admin')->middleware('auth')->group(function () {
     Route::delete('operadores/destroy/{id}',[OperadoresController::class, 'destroy'])->name('admin.operadores.destroy');
 
     // Manage Team
-    Route::get('entidades_formadores', [EntidadesFormadoreasController::class, 'index'])->name('admin.entidades_formadoreas');
-    Route::get('entidades_formadores/create', [EntidadesFormadoreasController::class, 'create'])->name('admin.entidades_formadoreas.create');
-    Route::post('entidades_formadores/create', [EntidadesFormadoreasController::class, 'store'])->name('admin.entidades_formadoreas.store');
-    Route::get('entidades_formadores/edit/{id}', [EntidadesFormadoreasController::class, 'edit'])->name('admin.entidades_formadoreas.edit');
-    Route::post('entidades_formadores/edit/{id}', [EntidadesFormadoreasController::class, 'update'])->name('admin.entidades_formadoreas.update');
-    Route::delete('entidades_formadores/destroy/{id}',[EntidadesFormadoreasController::class, 'destroy'])->name('admin.entidades_formadoreas.destroy');
+    Route::get('entidades_formadores', [EntidadesFormadoreasController::class, 'index'])->middleware('can:isAdminOrResponsable')->name('admin.entidades_formadoreas');
+    Route::get('entidades_formadores/create', [EntidadesFormadoreasController::class, 'create'])->middleware('can:isAdminOrResponsable')->name('admin.entidades_formadoreas.create');
+    Route::post('entidades_formadores/create', [EntidadesFormadoreasController::class, 'store'])->middleware('can:isAdminOrResponsable')->name('admin.entidades_formadoreas.store');
+    Route::get('entidades_formadores/edit/{id}', [EntidadesFormadoreasController::class, 'edit'])->middleware('can:isAdminOrResponsable')->name('admin.entidades_formadoreas.edit');
+    Route::post('entidades_formadores/edit/{id}', [EntidadesFormadoreasController::class, 'update'])->middleware('can:isAdminOrResponsable')->name('admin.entidades_formadoreas.update');
+    Route::delete('entidades_formadores/destroy/{id}',[EntidadesFormadoreasController::class, 'destroy'])->middleware('can:isAdmin')->name('admin.entidades_formadoreas.destroy');
 
      // Manage Admin
-     Route::get('users', [UserController::class, 'index'])->name('admin.user');
-     Route::post('users/{id}', [UserController::class, 'changepassword'])->name('admin.user.changepassword');
-     Route::get('users/create', [UserController::class, 'create'])->name('admin.user.create');
-     Route::post('users/create', [UserController::class, 'store'])->name('admin.user.store');
-     Route::get('users/edit/{id}', [UserController::class, 'edit'])->name('admin.user.edit');
-     Route::post('users/edit/{id}', [UserController::class, 'update'])->name('admin.user.update');
-     Route::delete('users/destroy/{id}',[UserController::class, 'destroy'])->name('admin.user.destroy');
+     Route::get('users', [UserController::class, 'index'])->middleware('can:isAdminOrResponsable')->name('admin.user');
+     Route::post('users/{id}', [UserController::class, 'changepassword'])->middleware('can:isAdminOrResponsable')->name('admin.user.changepassword');
+     Route::get('users/create', [UserController::class, 'create'])->middleware('can:isAdminOrResponsable')->name('admin.user.create');
+     Route::post('users/create', [UserController::class, 'store'])->middleware('can:isAdminOrResponsable')->name('admin.user.store');
+     Route::get('users/edit/{id}', [UserController::class, 'edit'])->middleware('can:isAdminOrResponsable')->name('admin.user.edit');
+     Route::post('users/edit/{id}', [UserController::class, 'update'])->middleware('can:isAdminOrResponsable')->name('admin.user.update');
+     Route::delete('users/destroy/{id}',[UserController::class, 'destroy'])->middleware('can:isAdmin')->name('admin.user.destroy');
 });
